Separate MongoDbConfig for tests from ApplicationTestConfig | then rename ApplicationTestConfig --> ApplicationTest
