+++ src/test/java/org/utn/marvellator/ApplicationTest.java
@@ -14,40 +14,20 @@ import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.test.SpringApplicationConfiguration;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
 import org.springframework.data.mongodb.config.AbstractMongoConfiguration;
 import org.springframework.data.mongodb.repository.config.EnableMongoRepositories;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 import org.springframework.test.context.web.WebAppConfiguration;
+import org.utn.marvellator.config.MongoDbConfig;
 import org.utn.marvellator.model.User;
 import org.utn.marvellator.repository.UserRepository;
 
 @Configuration
 @EnableAutoConfiguration
 @ComponentScan({"org.utn.marvellator.repository", "org.utn.marvellator.service"})
-public class ApplicationTestConfig {
-
-	@Rule
-	public FongoRule fongoRule = new FongoRule();
-
-	@Configuration
-//	@EnableMongoRepositories
-	static class MongoConfiguration extends AbstractMongoConfiguration {
-
-		@Override
-		protected String getDatabaseName() {
-			return "test-db";
-		}
-
-		@Override
-		public Mongo mongo() {
-			return new Fongo("test").getMongo();
-		}
-
-		@Override
-		protected String getMappingBasePackage() {
-			return "org.utn.marvellator";
-		}
-	}
+@Import(MongoDbConfig.class)
+public class ApplicationTest {
 
 	@Test
 	public void contextLoads() {
