+++ src/test/java/org/utn/marvellator/repository/UserRepositoryTest.java
@@ -7,7 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.SpringApplicationConfiguration;
 import org.springframework.dao.DuplicateKeyException;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
-import org.utn.marvellator.ApplicationTestConfig;
+import org.utn.marvellator.ApplicationTest;
 import org.utn.marvellator.model.User;
 
 
@@ -15,7 +15,7 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 
 @RunWith(SpringJUnit4ClassRunner.class)
-@SpringApplicationConfiguration(classes = ApplicationTestConfig.class)
+@SpringApplicationConfiguration(classes = ApplicationTest.class)
 public class UserRepositoryTest {
 
     @Autowired
