+++ src/test/java/org/utn/marvellator/service/UserServiceTest.java
@@ -5,14 +5,14 @@ import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.SpringApplicationConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
-import org.utn.marvellator.ApplicationTestConfig;
+import org.utn.marvellator.ApplicationTest;
 import org.utn.marvellator.model.User;
 import org.utn.marvellator.repository.UserRepository;
 
 import static org.junit.Assert.assertEquals;
 
 @RunWith(SpringJUnit4ClassRunner.class)
-@SpringApplicationConfiguration(classes = ApplicationTestConfig.class)
+@SpringApplicationConfiguration(classes = ApplicationTest.class)
 public class UserServiceTest {
 
     @Autowired
